requirements.txt
google-cloud-bigquery>=3.20.0
requests>=2.31.0
python-dotenv>=1.0.0
